+++ tests/test_core.py
@@ -4,6 +4,7 @@ except ImportError:
     pass
 
 from transitions import Machine, State, MachineError
+from transitions.core import listify
 from unittest import TestCase
 from .utils import Stuff, InheritedStuff
 
@@ -40,6 +41,12 @@ class TestTransitions(TestCase):
         m = Machine(model=s, states=states, transitions=transitions, initial='State2')
         s.advance()
         self.assertEquals(s.message, 'Hello World!')
+
+    def test_listify(self):
+        self.assertEquals(listify(4), [4])
+        self.assertEquals(listify(None), [])
+        self.assertEquals(listify((4, 5)), (4, 5))
+        self.assertEquals(listify([1, 3]), [1, 3])
 
     def test_property_initial(self):
         states = ['A', 'B', 'C', 'D']
@@ -150,6 +157,35 @@ class TestTransitions(TestCase):
         self.assertEquals(m.model.level, 2)
         m.model.move()
         self.assertEquals(m.model.level, 3)
+
+    def test_prepare(self):
+        m = Machine(Stuff(), states=['A', 'B', 'C'], initial='A')
+        m.add_transition('move', 'A', 'B', prepare='increase_level')
+        m.add_transition('move', 'B', 'C', prepare='increase_level')
+        m.add_transition('move', 'C', 'A', prepare='increase_level', conditions='this_fails')
+        m.add_transition('dont_move', 'A', 'C', prepare='increase_level')
+
+        m.prepare_move('increase_level')
+
+        m.model.move()
+        self.assertEquals(m.model.state, 'B')
+        self.assertEquals(m.model.level, 3)
+
+        m.model.move()
+        self.assertEquals(m.model.state, 'C')
+        self.assertEquals(m.model.level, 5)
+
+        # State does not advance, but increase_level still runs
+        m.model.move()
+        self.assertEquals(m.model.state, 'C')
+        self.assertEquals(m.model.level, 7)
+
+        # An invalid transition shouldn't execute the callback
+        with self.assertRaises(MachineError):
+            m.model.dont_move()
+
+        self.assertEquals(m.model.state, 'C')
+        self.assertEquals(m.model.level, 7)
 
     def test_state_model_change_listeners(self):
         s = self.stuff
@@ -366,3 +402,130 @@ class TestTransitions(TestCase):
         m = Machine(states=states, transitions=transitions, initial='A', async=True)
         m.walk(machine=m)
         self.assertEqual(m.current_state.name, 'C')
+
+    def test___getattr___and_identify_callback(self):
+        m = Machine(Stuff(), states=['A', 'B', 'C'], initial='A')
+        m.add_transition('move', 'A', 'B')
+        m.add_transition('move', 'B', 'C')
+
+        callback = m.__getattr__('before_move')
+        self.assertTrue(callable(callback))
+
+        with self.assertRaises(MachineError):
+            m.__getattr__('before_no_such_transition')
+
+        with self.assertRaises(MachineError):
+            m.__getattr__('before_no_such_transition')
+
+        with self.assertRaises(AttributeError):
+            m.__getattr__('__no_such_method__')
+
+        with self.assertRaises(AttributeError):
+            m.__getattr__('')
+
+        type, target = m._identify_callback('on_exit_foobar')
+        self.assertEqual(type, 'on_exit')
+        self.assertEqual(target, 'foobar')
+
+        type, target = m._identify_callback('on_exitfoobar')
+        self.assertEqual(type, None)
+        self.assertEqual(target, None)
+
+        type, target = m._identify_callback('notacallback_foobar')
+        self.assertEqual(type, None)
+        self.assertEqual(target, None)
+
+        type, target = m._identify_callback('totallyinvalid')
+        self.assertEqual(type, None)
+        self.assertEqual(target, None)
+
+        type, target = m._identify_callback('before__foobar')
+        self.assertEqual(type, 'before')
+        self.assertEqual(target, '_foobar')
+
+        type, target = m._identify_callback('before__this__user__likes__underscores___')
+        self.assertEqual(type, 'before')
+        self.assertEqual(target, '_this__user__likes__underscores___')
+
+        type, target = m._identify_callback('before_stuff')
+        self.assertEqual(type, 'before')
+        self.assertEqual(target, 'stuff')
+
+        type, target = m._identify_callback('before_trailing_underscore_')
+        self.assertEqual(type, 'before')
+        self.assertEqual(target, 'trailing_underscore_')
+
+        type, target = m._identify_callback('before_')
+        self.assertIs(type, None)
+        self.assertIs(target, None)
+
+        type, target = m._identify_callback('__')
+        self.assertIs(type, None)
+        self.assertIs(target, None)
+
+        type, target = m._identify_callback('')
+        self.assertIs(type, None)
+        self.assertIs(target, None)
+
+    def test_state_and_transition_with_underscore(self):
+        m = Machine(Stuff(), states=['_A_', '_B_', '_C_'], initial='_A_')
+        m.add_transition('_move_', '_A_', '_B_', prepare='increase_level')
+        m.add_transition('_after_', '_B_', '_C_', prepare='increase_level')
+        m.add_transition('_on_exit_', '_C_', '_A_', prepare='increase_level', conditions='this_fails')
+
+        m.model._move_()
+        self.assertEquals(m.model.state, '_B_')
+        self.assertEquals(m.model.level, 2)
+
+        m.model._after_()
+        self.assertEquals(m.model.state, '_C_')
+        self.assertEquals(m.model.level, 3)
+
+        # State does not advance, but increase_level still runs
+        m.model._on_exit_()
+        self.assertEquals(m.model.state, '_C_')
+        self.assertEquals(m.model.level, 4)
+
+    def test_callback_identification(self):
+        m = Machine(Stuff(), states=['A', 'B', 'C', 'D', 'E', 'F'], initial='A')
+        m.add_transition('transition', 'A', 'B', before='increase_level')
+        m.add_transition('after', 'B', 'C', before='increase_level')
+        m.add_transition('on_exit_A', 'C', 'D', before='increase_level', conditions='this_fails')
+        m.add_transition('check', 'C', 'E', before='increase_level')
+        m.add_transition('prepare', 'E', 'F', before='increase_level')
+        m.add_transition('before', 'F', 'A', before='increase_level')
+
+        m.before_transition('increase_level')
+        m.before_after('increase_level')
+        m.before_on_exit_A('increase_level')
+        m.after_check('increase_level')
+        m.before_prepare('increase_level')
+        m.before_before('increase_level')
+
+        m.model.transition()
+        self.assertEquals(m.model.state, 'B')
+        self.assertEquals(m.model.level, 3)
+
+        m.model.after()
+        self.assertEquals(m.model.state, 'C')
+        self.assertEquals(m.model.level, 5)
+
+        m.model.on_exit_A()
+        self.assertEquals(m.model.state, 'C')
+        self.assertEquals(m.model.level, 5)
+
+        m.model.check()
+        self.assertEquals(m.model.state, 'E')
+        self.assertEquals(m.model.level, 7)
+
+        m.model.prepare()
+        self.assertEquals(m.model.state, 'F')
+        self.assertEquals(m.model.level, 9)
+
+        m.model.before()
+        self.assertEquals(m.model.state, 'A')
+        self.assertEquals(m.model.level, 11)
+
+        # An invalid transition shouldn't execute the callback
+        with self.assertRaises(MachineError):
+                m.model.on_exit_A()