+++ Changelog.md
@@ -2,8 +2,8 @@
 
 ## 0.4.0 (April, 2016)
 Release 0.4 includes several new features:
-- New `name` Machine keyword customizes transitions logger output for easier debugging while running multiple instances
-- New `prepare` Transition keyword for callbacks before any 'conditions' are checked  (thanks to @TheMysteriousX)
+- New `name` Machine keyword customizes transitions logger output for easier debugging of multiple running instances
+- New `prepare` Transition keyword for callbacks before any 'conditions' are checked (thanks to @TheMysteriousX)
 - New `async` Machine keyword allows queueing of transitions (thanks to @khigia)
 - Nesting now supports custom (unicode) substate separators
 - Factory for convenient extension mixins